requirements.txt
absl-py==1.4.0
astor==0.8.1
backports.functools-lru-cache==1.6.4
bleach==6.0.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.12
cplex==12.8.0.0
cycler==0.11.0
Cython==0.29.14
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
docloud==1.0.375
docplex==2.25.236
entrypoints==0.4
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
idna==3.4
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==6.2.0
ipython-genutils==0.2.0
ipywidgets==7.3.0
jedi==0.18.2
Jinja2==2.11.3
joblib==1.1.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==5.2.0
jupyter-core==4.9.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
nb-conda-kernels==2.3.1
nbconvert==5.6.1
nbformat==5.1.3
nest-asyncio==1.5.6
notebook==5.7.11
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==0.25.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.17.0
prompt-toolkit==1.0.15
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.12.0
pyjnius==1.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==19.0.2
qtconsole==4.4.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.8.2
simplegeneric==0.8.1
six==1.16.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.6.0
torch==1.8.0
tornado==6.1
traitlets==4.3.3
typing-extensions==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.3.1
wrapt==1.15.0
zipp==3.6.0